Added more file separation to make program less crowded.
